Move contract deploy helpers to setup.rs.
