+++ tests/integration/src/xvm.rs
@@ -20,10 +20,7 @@ use crate::setup::*;
 
 use astar_primitives::xvm::{Context, VmId, XvmCall};
 use frame_support::{traits::Currency, weights::Weight};
-use pallet_contracts::{CollectEvents, DebugInfo};
-use pallet_contracts_primitives::Code;
 use parity_scale_codec::Encode;
-use sp_core::{H160, H256, U256};
 use sp_runtime::MultiAddress;
 
 /*
@@ -50,14 +47,6 @@ contract Payable {
 
  */
 const EVM_PAYABLE: &str = "6080604052336000806101000a81548173ffffffffffffffffffffffffffffffffffffffff021916908373ffffffffffffffffffffffffffffffffffffffff1602179055506102d6806100536000396000f3fe6080604052600436106100345760003560e01c80633ccfd60b146100395780638da5cb5b14610050578063d0e30db01461007b575b600080fd5b34801561004557600080fd5b5061004e610085565b005b34801561005c57600080fd5b5061006561015b565b60405161007291906101c2565b60405180910390f35b61008361017f565b005b600047905060008060009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16826040516100d19061020e565b60006040518083038185875af1925050503d806000811461010e576040519150601f19603f3d011682016040523d82523d6000602084013e610113565b606091505b5050905080610157576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040161014e90610280565b60405180910390fd5b5050565b60008054906101000a900473ffffffffffffffffffffffffffffffffffffffff1681565b565b600073ffffffffffffffffffffffffffffffffffffffff82169050919050565b60006101ac82610181565b9050919050565b6101bc816101a1565b82525050565b60006020820190506101d760008301846101b3565b92915050565b600081905092915050565b50565b60006101f86000836101dd565b9150610203826101e8565b600082019050919050565b6000610219826101eb565b9150819050919050565b600082825260208201905092915050565b7f4661696c656420746f2077697468647261772045746865720000000000000000600082015250565b600061026a601883610223565b915061027582610234565b602082019050919050565b600060208201905081810360008301526102998161025d565b905091905056fea2646970667358221220bd8883b6a524d12ac9c29f105fdd1a0221a0436a79002f2a04e69d252596a62a64736f6c63430008120033";
-const EVM_PAYABLE_ADDR: &str = "4a15f25194b60fd07860b3a20a060fd003cae5a6";
-
-fn evm_payable_addr() -> Vec<u8> {
-    hex::decode(EVM_PAYABLE_ADDR).unwrap()
-}
-fn evm_payable_addr_h160() -> H160 {
-    H160::from_slice(evm_payable_addr().as_slice())
-}
 
 /* WASM payable:
 
@@ -104,14 +93,6 @@ contract CallXVMPayble {
 
  */
 const CALL_WASM_PAYBLE: &str = "608060405234801561001057600080fd5b506105e6806100206000396000f3fe60806040526004361061001e5760003560e01c80634012b91414610023575b600080fd5b61003d600480360381019061003891906101a3565b610054565b60405161004b9291906102e3565b60405180910390f35b6000606061500573ffffffffffffffffffffffffffffffffffffffff1663e5d9bac0601f89898989896040518763ffffffff1660e01b815260040161009e969594939291906103b0565b6000604051808303816000875af11580156100bd573d6000803e3d6000fd5b505050506040513d6000823e3d601f19601f820116820180604052508101906100e69190610554565b915091509550959350505050565b6000604051905090565b600080fd5b600080fd5b600080fd5b600080fd5b600080fd5b60008083601f84011261012d5761012c610108565b5b8235905067ffffffffffffffff81111561014a5761014961010d565b5b60208301915083600182028301111561016657610165610112565b5b9250929050565b6000819050919050565b6101808161016d565b811461018b57600080fd5b50565b60008135905061019d81610177565b92915050565b6000806000806000606086880312156101bf576101be6100fe565b5b600086013567ffffffffffffffff8111156101dd576101dc610103565b5b6101e988828901610117565b9550955050602086013567ffffffffffffffff81111561020c5761020b610103565b5b61021888828901610117565b9350935050604061022b8882890161018e565b9150509295509295909350565b60008115159050919050565b61024d81610238565b82525050565b600081519050919050565b600082825260208201905092915050565b60005b8381101561028d578082015181840152602081019050610272565b60008484015250505050565b6000601f19601f8301169050919050565b60006102b582610253565b6102bf818561025e565b93506102cf81856020860161026f565b6102d881610299565b840191505092915050565b60006040820190506102f86000830185610244565b818103602083015261030a81846102aa565b90509392505050565b6000819050919050565b600060ff82169050919050565b6000819050919050565b600061034f61034a61034584610313565b61032a565b61031d565b9050919050565b61035f81610334565b82525050565b82818337600083830152505050565b6000610380838561025e565b935061038d838584610365565b61039683610299565b840190509392505050565b6103aa8161016d565b82525050565b60006080820190506103c56000830189610356565b81810360208301526103d8818789610374565b905081810360408301526103ed818587610374565b90506103fc60608301846103a1565b979650505050505050565b61041081610238565b811461041b57600080fd5b50565b60008151905061042d81610407565b92915050565b600080fd5b7f4e487b7100000000000000000000000000000000000000000000000000000000600052604160045260246000fd5b61047082610299565b810181811067ffffffffffffffff8211171561048f5761048e610438565b5b80604052505050565b60006104a26100f4565b90506104ae8282610467565b919050565b600067ffffffffffffffff8211156104ce576104cd610438565b5b6104d782610299565b9050602081019050919050565b60006104f76104f2846104b3565b610498565b90508281526020810184848401111561051357610512610433565b5b61051e84828561026f565b509392505050565b600082601f83011261053b5761053a610108565b5b815161054b8482602086016104e4565b91505092915050565b6000806040838503121561056b5761056a6100fe565b5b60006105798582860161041e565b925050602083015167ffffffffffffffff81111561059a57610599610103565b5b6105a685828601610526565b915050925092905056fea264697066735822122047908cecfa9ace275a4ba96e787bb0d1541ec599c370983693c6cd9c1f5b7dbe64736f6c63430008120033";
-const CALL_WASM_PAYBLE_ADDR: &str = "79b3ca9f410de3d98c1dafac0e291871f752bc7a";
-
-fn call_wasm_payable_addr() -> Vec<u8> {
-    hex::decode(CALL_WASM_PAYBLE_ADDR).unwrap()
-}
-fn call_wasm_payable_addr_h160() -> H160 {
-    H160::from_slice(call_wasm_payable_addr().as_slice())
-}
 
 /* Call EVM Payable:
 
@@ -171,50 +152,10 @@ impl Environment for CustomEnvironment {
 
  */
 
-fn deploy_evm_contract(code: &str, assert_addr: H160) {
-    assert_ok!(EVM::create2(
-        RuntimeOrigin::root(),
-        alith(),
-        hex::decode(code).unwrap(),
-        H256::zero(),
-        U256::zero(),
-        1_000_000,
-        U256::from(DefaultBaseFeePerGas::get()),
-        None,
-        None,
-        vec![],
-    ));
-    System::assert_last_event(RuntimeEvent::EVM(pallet_evm::Event::Created {
-        address: assert_addr,
-    }));
-}
-
-fn deploy_wasm_contract(name: &str) -> AccountId32 {
-    let path = format!("resource/{}.wasm", name);
-    let code = std::fs::read(path).unwrap();
-    let instantiate_result = Contracts::bare_instantiate(
-        ALICE,
-        0,
-        Weight::from_parts(10_000_000_000, 1024 * 1024),
-        None,
-        Code::Upload(code),
-        // `new` constructor
-        hex::decode("9bae9d5e").unwrap(),
-        vec![],
-        DebugInfo::Skip,
-        CollectEvents::Skip,
-    );
-
-    let address = instantiate_result.result.unwrap().account_id;
-    // On instantiation, the contract got existential deposit.
-    assert_eq!(Balances::free_balance(&address), ExistentialDeposit::get(),);
-    address
-}
-
 #[test]
 fn evm_payable_call_via_xvm_works() {
     new_test_ext().execute_with(|| {
-        deploy_evm_contract(EVM_PAYABLE, evm_payable_addr_h160());
+        let evm_payable_addr = deploy_evm_contract(EVM_PAYABLE);
 
         let value = 1_000_000_000;
         assert_ok!(Xvm::call(
@@ -224,13 +165,13 @@ fn evm_payable_call_via_xvm_works() {
             },
             VmId::Evm,
             ALICE,
-            evm_payable_addr(),
+            evm_payable_addr.as_ref().to_vec(),
             // Calling `deposit`
             hex::decode("d0e30db0").unwrap(),
             value,
         ));
         assert_eq!(
-            Balances::free_balance(account_id_from(evm_payable_addr_h160())),
+            Balances::free_balance(account_id_from(evm_payable_addr)),
             value
         );
 
@@ -241,13 +182,13 @@ fn evm_payable_call_via_xvm_works() {
             },
             VmId::Evm,
             ALICE,
-            evm_payable_addr(),
+            evm_payable_addr.as_ref().to_vec(),
             // `Calling withdraw`
             hex::decode("3ccfd60b").unwrap(),
             0,
         ));
         assert_eq!(
-            Balances::free_balance(account_id_from(evm_payable_addr_h160())),
+            Balances::free_balance(account_id_from(evm_payable_addr)),
             ExistentialDeposit::get(),
         );
     });
@@ -279,13 +220,13 @@ fn wasm_payable_call_via_xvm_works() {
 fn calling_wasm_payable_from_evm_works() {
     new_test_ext().execute_with(|| {
         let wasm_payable_addr = deploy_wasm_contract("payable");
-        deploy_evm_contract(CALL_WASM_PAYBLE, call_wasm_payable_addr_h160());
+        let call_wasm_payable_addr = deploy_evm_contract(CALL_WASM_PAYBLE);
 
         let value = 1_000_000_000;
         assert_ok!(EVM::call(
             RuntimeOrigin::root(),
             alith(),
-            call_wasm_payable_addr_h160(),
+            call_wasm_payable_addr.clone(),
             // to: 0x00a8f69d59df362b69a8d4acdb9001eb3e1b8d067b8fdaa70081aed945bde5c48c
             // input: 0x0000002a (deposit)
             // value: 1000000000
@@ -304,7 +245,7 @@ fn calling_wasm_payable_from_evm_works() {
             value,
         );
         assert_eq!(
-            Balances::free_balance(&account_id_from(call_wasm_payable_addr_h160())),
+            Balances::free_balance(&account_id_from(call_wasm_payable_addr)),
             ExistentialDeposit::get(),
         );
     });
@@ -313,7 +254,7 @@ fn calling_wasm_payable_from_evm_works() {
 #[test]
 fn calling_evm_payable_from_wasm_works() {
     new_test_ext().execute_with(|| {
-        deploy_evm_contract(EVM_PAYABLE, evm_payable_addr_h160());
+        let evm_payable_addr = deploy_evm_contract(EVM_PAYABLE);
         let wasm_address = deploy_wasm_contract("call_xvm_payable");
 
         let value = UNIT;
@@ -324,7 +265,7 @@ fn calling_evm_payable_from_wasm_works() {
         let mock_unified_wasm_account = account_id_from(h160_from(wasm_address.clone()));
         let _ = Balances::deposit_creating(&mock_unified_wasm_account, value);
 
-        let evm_payable = evm_payable_addr_h160().as_ref().to_vec();
+        let evm_payable = evm_payable_addr.as_ref().to_vec();
         let deposit_func = hex::decode("d0e30db0").unwrap();
         let input = hex::decode("0000002a")
             .unwrap()
@@ -343,7 +284,7 @@ fn calling_evm_payable_from_wasm_works() {
         ));
 
         assert_eq!(
-            Balances::free_balance(account_id_from(evm_payable_addr_h160())),
+            Balances::free_balance(account_id_from(evm_payable_addr)),
             value
         );
